Optimization offers rerendering when hover
