+++ project/src/pages/main-page/main-page.tsx
@@ -1,38 +1,26 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import Layout from '../../components/layout/layout';
-import Map from '../../components/map/map';
-import OffersList from '../../components/offers-list/offers-list';
-import Sort from '../../components/sort/sort';
 import Tabs from '../../components/tabs/tabs';
-import { OfferCardType } from '../../const';
 import { useAppSelector } from '../../hooks/use-app-selector/use-app-selector';
 import cn from 'classnames';
-import EmptyMessage from '../../components/empty-message/empty-message';
 import Loader from '../../components/loader/loader';
-import { getOffersStatus, getRenderedOffers } from '../../store/reducers/offers/selectors';
-import { getCurrentCity, getCurrentSortType } from '../../store/reducers/app/selectors';
+import { getIsEmpty, getOffersStatus } from '../../store/reducers/offers/selectors';
+import { getCurrentCity } from '../../store/reducers/app/selectors';
 import { useAppDispatch } from '../../hooks/use-app-dispatch/use-app-dispatch';
 import { fetchOffersAction } from '../../store/api-actions';
 import ErrorPage from '../error-page/error-page';
+import OffersBoard from '../../components/offers-board/offers-board';
 
 function MainPage(): JSX.Element {
   const status = useAppSelector(getOffersStatus);
   const city = useAppSelector(getCurrentCity);
-  const sortType = useAppSelector(getCurrentSortType);
-  const renderedOffers = useAppSelector(getRenderedOffers);
+  const isEmpty = useAppSelector(getIsEmpty);
 
-  const isEmpty = !renderedOffers.length;
-
-  const [selectedOfferId, setSelectedOfferId] = useState<number | null>(null);
   const dispatch = useAppDispatch();
 
   useEffect(() => {
     dispatch(fetchOffersAction());
   }, [dispatch]);
-
-  const onCardHover = (offerId: number | null): void => {
-    setSelectedOfferId(offerId);
-  };
 
   if (status.isLoading) {
     return <Loader isSmall={false} />;
@@ -49,36 +37,8 @@ function MainPage(): JSX.Element {
 
         <Tabs currentCity={city} />
 
-        <div className="cities">
-          <div className={cn('cities__places-container container', isEmpty && 'cities__places-container--empty')}>
-            {isEmpty ?
-              <EmptyMessage currentCity={city} />
-              : (
-                <section className="cities__places places">
-                  <h2 className="visually-hidden">Places</h2>
-                  <b className="places__found">{renderedOffers.length} places to stay in {city}</b>
+        <OffersBoard />
 
-                  <Sort currentSortType={sortType} />
-
-                  <OffersList
-                    offers={renderedOffers}
-                    classNames="cities__places-list places__list"
-                    offerCardType={OfferCardType.Main}
-                    onCardHover={onCardHover}
-                  />
-                </section>)}
-
-            <div className="cities__right-section">
-              {!isEmpty && (
-                <Map
-                  className="cities"
-                  location={renderedOffers[0].city.location}
-                  offers={renderedOffers}
-                  selectedOfferId={selectedOfferId}
-                />)}
-            </div>
-          </div>
-        </div>
       </main>
     </Layout >
   );
